+++ glue_ginga/qt/layer_artist.py
